requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
schedule==1.2.0
loguru==0.7.2
sqlalchemy==2.0.23
alembic==1.13.1 